+++ prj_Explorer/src/main/java/com/explorer/repos/UtenteDAO.java
@@ -10,7 +10,7 @@ import org.springframework.data.repository.query.Param;
 import com.explorer.entities.Utente;
 
 
-public interface UtenteDAO extends JpaRepository<Utente,Long> {
+public interface UtenteDAO extends JpaRepository<Utente,Integer> {
 	
 	@Query("SELECT u FROM Utente u WHERE u.nome = :nome")
 	List<Utente> findByNome(@Param("nome") String nome);
